refactor(index): refactor to depend
